start game models

+++ dojo-starter/src/models/position.cairo
@@ -9,8 +9,8 @@ struct Position {
 
 #[derive(Copy, Drop, Serde, Introspect)]
 struct Vec2 {
-    x: u32,
-    y: u32
+    x: u8,
+    y: u8
 }
 
 trait Vec2Trait {
